move the class to another file
